port building hours to tanstack-query
